Move index.html generation out of frontend generator module
